+++ src/main/java/seedu/address/logic/parser/ImportCommandParser.java
@@ -1,0 +1,8 @@
+package seedu.address.logic.parser;
+
+/**
+ * Parses input arguments and creates a new ImportCommand object.
+ */
+public class ImportCommandParser {
+
+}
