+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.Test;
 
 import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.parser.person.DeleteCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
