+++ src/test/java/seedu/address/logic/parser/DeleteEventParserTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.Test;
 
 import seedu.address.logic.commands.DeleteEventCommand;
+import seedu.address.logic.parser.event.DeleteEventParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
