+++ src/test/java/seedu/address/logic/parser/EditEventCommandParserTest.java
@@ -24,6 +24,7 @@ import org.junit.Test;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditEventCommand;
 import seedu.address.logic.commands.EditEventCommand.EditEventDescriptor;
+import seedu.address.logic.parser.event.EditEventParser;
 import seedu.address.model.property.PropertyManager;
 import seedu.address.testutil.EditEventDescriptorBuilder;
