+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import org.junit.Test;
 
 import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.parser.person.FindCommandParser;
 import seedu.address.model.property.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
