+++ src/test/java/seedu/address/logic/parser/SelectCommandParserTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.Test;
 
 import seedu.address.logic.commands.SelectCommand;
+import seedu.address.logic.parser.person.SelectCommandParser;
 
 /**
  * Test scope: similar to {@code DeleteCommandParserTest}.
